extract a bunch of stuff to nvrh_ssh
